inventario-app/backend/app/api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
gunicorn
tensorflow
numpy
supabase-py